services/telegram-bot/requirements.txt
python-telegram-bot>=20.0
confluent-kafka
python-dotenv
